+++ instrumentation/agent/src/main/kotlin/ApplicationCodeTransformer.kt
@@ -45,6 +45,7 @@ class ApplicationCodeTransformer : ClassFileTransformer {
             "com.github.ajalt",
         ) ||
         (dotClassName.startsWith("org.pastalab.fray") &&
+            !dotClassName.startsWith("org.pastalab.fray.example") &&
             !dotClassName.startsWith("org.pastalab.fray.benchmark") &&
             !dotClassName.startsWith("org.pastalab.fray.test") &&
             !dotClassName.startsWith("org.pastalab.fray.junit.internal") &&
@@ -78,7 +79,7 @@ class ApplicationCodeTransformer : ClassFileTransformer {
       cv = TimeInstrumenter(cv)
       cv = SkipMethodInstrumenter(cv)
       cv = ObjectHashCodeInstrumenter(cv, false)
-      cv = AtomicGetInstrumenter(cv)
+      //      cv = AtomicGetInstrumenter(cv)
       //      cv = ToStringInstrumenter(cv)
       val classVersionInstrumenter = ClassVersionInstrumenter(cv)
       cv = ArrayOperationInstrumenter(classVersionInstrumenter)