+++ integration-test/src/test/java/org/pastalab/fray/test/FrayTestCase.java
@@ -13,19 +13,7 @@ import org.pastalab.fray.core.command.MethodExecutor;
 import org.pastalab.fray.core.randomness.ControlledRandom;
 import org.pastalab.fray.core.scheduler.PCTScheduler;
 import org.pastalab.fray.core.scheduler.POSScheduler;
-import org.pastalab.fray.core.scheduler.RandomScheduler;
-import org.pastalab.fray.test.fail.cdl.CountDownLatchDeadlockUnblockMultiThread;
-import org.pastalab.fray.test.fail.park.ParkDeadlock;
-import org.pastalab.fray.test.fail.rwlock.ReentrantReadWriteLockDeadlock;
 import org.pastalab.fray.test.fail.thread.ThreadExitDeadlock;
-import org.pastalab.fray.test.fail.wait.NotifyOrder;
-import org.pastalab.fray.test.success.condition.ConditionAwaitTimeoutNotifyInterrupt;
-import org.pastalab.fray.test.success.rwlock.ReentrantReadWriteLockDowngradingNoDeadlock;
-import org.pastalab.fray.test.success.rwlock.ReentrantReadWriteLockNoDeadlock;
-import org.pastalab.fray.test.success.rwlock.ReentrantReadWriteLockNormalLockUnlock;
-import org.pastalab.fray.test.success.stampedlock.StampedLockTryLockNoDeadlock;
-import org.pastalab.fray.test.success.thread.ThreadInterruptionWithoutStart;
-import org.pastalab.fray.test.success.time.TestTime;
 
 import java.util.*;