+++ rmi/src/main/kotlin/org/pastalab/fray/rmi/Constant.kt
@@ -1,0 +1,5 @@
+package org.pastalab.fray.rmi
+
+object Constant {
+  const val REGISTRY_PORT = 15213
+}