+++ CHANGELOG.md
@@ -4,9 +4,15 @@
 
 ### Added
 
+- Allow the Fray debugger to highlight all lines of a thread that is being executed.
+- Enable click action in the thread stack trace to navigate to the source code.
+- Add translated Deadlock Empire games!
+- Change the focus of the editor when a new thread is selected in the Fray debugger.
+
 ### Changed
 
 - Simplify the `mainExit` logic and introduce `MainExiting` state.
+- Change the z order of highlighting to be on top of the editor.
 
 ### Deprecated